test: test a program with syntax error
