+++ test/cdlsolver_test.py
@@ -1,6 +1,7 @@
 import io
 
 from cdlsolver.control.control import Control
+from test.output_model import OutputModel, Output
 import unittest
 import logging
 from time import perf_counter as timer
@@ -88,65 +89,8 @@ DIR_AB_TEST_OUT = 'E:/Projects/cdlsolver/test/test_programs/direct_abnormal.out'
 DIR_SCHOOL_TEST = ['E:/Projects/cdlsolver/test/test_programs/school.lp']
 DIR_SCHOOL_TEST_OUT = 'E:/Projects/cdlsolver/test/test_programs/school.out'
 
-
-class OutputModel:
-    def __init__(self, x=None, y=None):
-        if x is None:
-            x = set()
-        if y is None:
-            y = set()
-        self._x = x
-        self._y = y
-
-    def load(self, str):
-        parts = str.strip().replace(' ', '').split(';')
-        self._x = set(parts[0].replace('x:', '').split(','))
-        self._y = set(parts[1].replace('y:', '').split(','))
-
-    def __key(self):
-        return ';'.join([','.join([str(x) for x in self._x]), ','.join([str(y) for y in self._y])])
-
-    def __hash__(self):
-        return hash(self.__key())
-
-    def __eq__(self, other):
-        if isinstance(other, OutputModel):
-            return self._x == other._x and self._y == other._y
-        return NotImplemented
-
-    def __repr__(self):
-        lis_x = [str(x) for x in self._x]
-        lis_y = [str(y) for y in self._y]
-        lis_x.sort()
-        lis_y.sort()
-        return "x:" + ','.join(lis_x) + ";y:" + ','.join(lis_y) + ";"
-
-
-class Output:
-    def __init__(self, models=None):
-        if models is None:
-            models = set()
-        self._models = models
-
-    def add(self, model):
-        self._models.add(model)
-
-    def load(self, str):
-        for line in str.splitlines():
-            if line.startswith('x'):
-                model = OutputModel()
-                model.load(line)
-                self._models.add(model)
-
-    def count(self):
-        return len(self._models)
-
-    def __eq__(self, other):
-        if isinstance(other, Output):
-            set_mod = set([str(m) for m in self._models])
-            set_other = set([str(m) for m in other._models])
-            return set_mod == set_other
-        return NotImplemented
+DIR_SYNERR_1_TEST = ['E:/Projects/cdlsolver/test/test_programs/naf_default.lp']
+DIR_SYNERR_1_TEST_OUT = 'E:/Projects/cdlsolver/test/test_programs/naf_default.out'
 
 
 class ControlTest(unittest.TestCase):
@@ -162,32 +106,40 @@ class ControlTest(unittest.TestCase):
         start = timer()
         control = Control()
 
-        for file_path in paths:
-            control.load(file_path)
+        try:
+            for file_path in paths:
+                control.load(file_path)
 
-        models = control.solve()
-        end = timer()
-        print(f'Elapsed time: {(end - start):.6f}s')
+            models = control.solve()
+            end = timer()
+            print(f'Elapsed time: {(end - start):.6f}s')
 
-        output = Output()
-        cnt = 0
-        for model in models:
-            cnt += 1
-            output_model = OutputModel()
-            output_model.load(str(model))
-            output.add(output_model)
-            print(f'default model: {cnt}\n{model}')
+            output = Output()
+            cnt = 0
+            for model in models:
+                cnt += 1
+                output_model = OutputModel()
+                output_model.load(str(model))
+                output.add(output_model)
+                print(f'default model: {cnt}\n{model}')
 
-        if output.count() == 0:
-            print('Unsatisfiable')
-        else:
-            print('Satisfiable')
+            if output.count() == 0:
+                print('Unsatisfiable')
+            else:
+                print('Satisfiable')
 
-        std_output = Output()
-        with open(out, 'r') as out_file:
-            std_output.load(out_file.read())
-            self.assertEqual(std_output, output)
-            out_file.close()
+            std_output = Output()
+            with open(out, 'r') as out_file:
+                std_output.load(out_file.read())
+                self.assertEqual(std_output, output)
+                out_file.close()
+        except SyntaxError as e:
+            print(e.msg)
+            with open(out, 'r') as out_file:
+                std_err = out_file.read()
+                self.assertEqual(std_err, e.msg)
+                out_file.close()
+
 
     def test_basic(self):
         # reading
@@ -273,3 +225,6 @@ class ControlTest(unittest.TestCase):
 
     def test_school(self):
         self._test_control(DIR_SCHOOL_TEST, DIR_SCHOOL_TEST_OUT)
+
+    def test_syntax_error_1(self):
+        self._test_control(DIR_SYNERR_1_TEST, DIR_SYNERR_1_TEST_OUT)
