+++ test/test_programs/naf_default.lp
@@ -1,0 +1,1 @@
+a :- &c{not b}.
